Use proper logger levels instead of println.
This should prevent debug spew on standard build configs.

Errors are used for cycle finder or translation issues.

A warning is used for the last message about lack of tasks to hook up j2objc to,
while debug level is used for all else.  Add --debug to your gradle or gradlew command
to see this messages.

Translation and test output is routed to debug, unless an error occurs, in which
case it is routed to error.

TESTED=yes