FAQ.md - improve wording & fix nuances
Note including build/* in your srcDir may not work as you intend; you need to have task dependencies between the things that generated those build/* files and.  I replaced it with an example that does not rely on such assumptions.

In the future, please include references to authorities - we don't want to become the documentation source-of-truth for much of this info, just a way to solve the easiest instances, and then point people in the right directions for details.