projectExec regex matching for better error on unexpected output
If the command line succeeds but the task regex match fails, e.g. can’t
find (OK XX tests) for example, then the error is missing the command
line, stdout or stderr. This change pushes the regex match in to
projectExec, so it can produce more meaningful error messages. This
duplicates the regex match done in the task itself but this is a small
tradeoff.

projectExec
- Intercept all exceptions so that they all get useful error messages
- matchRegexOutputsRequired parameter
- Distinguish between command line failure or regex failure

Tasks
- Rework exceptions and error messages to be more descriptive
- Reworked CycleFinderTask exception handling
- Escaped regex match examples

TESTED=added test: projectExec_MatchRegexFailed