Build external Java libs into standalone native libs.
Rename the old `j2objcTranslation` to `j2objcTranslationClosure`

Add a new dependency type `j2objcTranslation` that allows you to
convert an entire source jar into a standalone library as a separate Gradle project.
You then depend on that project instead of the external library when
you want to use the functionality elsewhere.

This causes the external library to be translated and compiled only once,
and not partially and duplicatively as with --build-closure.

Fixes #429; fixes #419

TESTED=yes