Customization of Xcode build configuration names
Adds two new configs, `xcodeDebugConfigurations` and `xcodeReleaseConfigurations`, to allow customization of the Xcode build configuration names that are specified in the Podfile as linking with the debug and release translated libs. These default to the previously hardcoded `['Debug']` and `['Release']`.

I, Daniel Dickison (danieldickison@gmail.com, https://github.com/danieldickison), certify that
a) this Contribution is my original work, and
b) I have the right to submit this Contribution under the Apache License,
   Version 2.0 (the "License") available at
   http://www.apache.org/licenses/LICENSE-2.0, and
c) I am submitting this Contribution under the License.